Create new google-cloud-translate gem package
