+++ google-cloud-translate/.gitignore
@@ -1,0 +1,9 @@
+Gemfile.lock
+coverage/*
+doc/*
+pkg/*
+html/*
+jsondoc/*
+
+# Ignore YARD stuffs
+.yardoc
