+++ google-cloud-translate/Gemfile
@@ -1,0 +1,6 @@
+source "http://rubygems.org"
+
+gemspec
+
+gem "rake"
+gem "google-cloud-core", path: "../google-cloud-core"
