+++ google-cloud-translate/google-cloud-translate.gemspec
@@ -1,0 +1,33 @@
+# -*- encoding: utf-8 -*-
+require File.expand_path("../lib/google/cloud/translate/version", __FILE__)
+
+Gem::Specification.new do |gem|
+  gem.name          = "google-cloud-translate"
+  gem.version       = Google::Cloud::Translate::VERSION
+
+  gem.authors       = ["Mike Moore", "Chris Smith"]
+  gem.email         = ["mike@blowmage.com", "quartzmo@gmail.com"]
+  gem.description   = "Gcloud is the official library for interacting with Google Cloud."
+  gem.summary       = "API Client library for Google Cloud"
+  gem.homepage      = "http://googlecloudplatform.github.io/gcloud-ruby/"
+  gem.license       = "Apache-2.0"
+
+  gem.files         = `git ls-files -- lib/*`.split("\n") +
+                      `git ls-files -- docs/*`.split("\n")
+  gem.require_paths = ["lib"]
+
+  gem.required_ruby_version = ">= 2.0.0"
+
+  gem.add_dependency "google-cloud-core"
+  gem.add_dependency "google-api-client", "~> 0.9.11"
+
+  gem.add_development_dependency "minitest", "~> 5.9"
+  gem.add_development_dependency "minitest-autotest", "~> 1.0"
+  gem.add_development_dependency "minitest-focus", "~> 1.1"
+  gem.add_development_dependency "minitest-rg", "~> 5.2"
+  gem.add_development_dependency "autotest-suffix", "~> 1.1"
+  gem.add_development_dependency "rubocop", "<= 0.35.1"
+  gem.add_development_dependency "simplecov", "~> 0.9"
+  gem.add_development_dependency "coveralls", "~> 0.7"
+  gem.add_development_dependency "yard", "~> 0.9"
+end
