+++ google-cloud-translate/lib/google/cloud/translate/detection.rb
