+++ google-cloud-translate/test/google/cloud/translate/detect_test.rb
