+++ google-cloud-translate/test/google/cloud/translate/languages_test.rb
