+++ google-cloud-translate/test/google/cloud/translate/translate_test.rb
