+++ google-cloud-translate/test/google/cloud/translate_test.rb
