+++ google-cloud-translate/test/helper.rb
@@ -1,0 +1,46 @@
+# Copyright 2016 Google Inc. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+gem "minitest"
+require "minitest/autorun"
+require "minitest/focus"
+require "minitest/rg"
+require "ostruct"
+require "json"
+require "base64"
+require "google/cloud/translate"
+
+##
+# Monkey-Patch Google API Client to support Mocks
+module Google::Apis::Core::Hashable
+  ##
+  # Minitest Mock depends on === to match same-value objects.
+  # By default, the Google API Client objects do not match with ===.
+  # Therefore, we must add this capability.
+  # This module seems like as good a place as any...
+  def === other
+    return(to_h === other.to_h) if other.respond_to? :to_h
+    super
+  end
+end
+
+class MockTranslate < Minitest::Spec
+  let(:key) { "test-api-key" }
+  let(:translate) { Google::Cloud::Translate::Api.new(Google::Cloud::Translate::Service.new(key)) }
+
+  # Register this spec type for when :mock_translate is used.
+  register_spec_type(self) do |desc, *addl|
+    addl.include? :mock_translate
+  end
+end
